move hashBlockchainId to util package
