+++ packages-internal/utils/src/index.ts
@@ -1,1 +1,2 @@
+export { hashBlockchainId } from './utils/hash-blockchain-id';
 export { TokenService } from './services/token-service/token-service';
