+++ packages-internal/utils/src/utils/hash-blockchain-id.ts
