+++ packages/avalanche-module/src/module.ts
@@ -16,7 +16,7 @@ import { getNetworkFee } from './handlers/get-network-fee';
 import { getTransactionHistory } from './handlers/get-transaction-history/get-transaction-history';
 import { getEnv } from './env';
 import { AvalancheGlacierService } from './services/glacier-service/glacier-service';
-import { hashBlockchainId } from './utils/hash-blockchain-id';
+import { hashBlockchainId } from '@internal/utils';
 
 export class AvalancheModule implements Module {
   #glacierService: AvalancheGlacierService;
